requirements-3.txt
streamlit
tensorflow
keras
numpy
matplotlib
pillow